requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
discord.py==1.4.1
gevent==20.5.0
greenlet==0.4.15
grequests==0.6.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pycparser==2.20
pylint==2.5.2
requests==2.23.0
six==1.14.0
soupsieve==2.0
toml==0.10.1
urllib3==1.25.9
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
